Easier config for federated pages
