+++ app/src/main/java/io/apicurio/registry/ui/beans/ConfigJs.java
@@ -24,7 +24,6 @@ import io.quarkus.runtime.annotations.RegisterForReflection;
 @RegisterForReflection
 public class ConfigJs {
 
-    public String mode;
     public ConfigJsArtifacts artifacts = new ConfigJsArtifacts();
     public ConfigJsUi ui = new ConfigJsUi();
     public ConfigJsFeatures features = new ConfigJsFeatures();
