+++ ui/config/config-keycloakjs.js
@@ -1,7 +1,5 @@
 var ApicurioRegistryConfig = {
-    mode: "dev",
     artifacts: {
-        type: "rest",
         url: "http://localhost:8080/apis/registry"
     },
     ui: {
