+++ ui/src/app/pages/artifacts/artifacts.federated.tsx
