+++ ui/src/services/config/config.type.ts
@@ -19,28 +19,39 @@ export interface FeaturesConfig {
     readOnly?: boolean;
     breadcrumbs?: boolean;
     multiTenant?: boolean;
-    multiTenantUrl?: string;
 }
 
 export interface ArtifactsConfig {
     url: string;
-    type: string;
 }
 
 export interface UiConfig {
-    contextPath: string|null;
-    navPrefixPath: string|null;
+    contextPath?: string;
+    navPrefixPath?: string;
 }
 
 export interface AuthConfig {
     type: string;
-    options: any;
+}
+
+// Used when `type=keycloakjs`
+export interface KeycloakJsAuthConfig extends AuthConfig {
+    options?: any;
+}
+
+// Used when `type=none`
+export interface NoneAuthConfig extends AuthConfig {
+
+}
+
+// Used when `type=gettoken`
+export interface GetTokenAuthConfig extends AuthConfig {
+    getToken: () => string;
 }
 
 export interface ConfigType {
-    mode: string;
     artifacts: ArtifactsConfig;
-    auth: AuthConfig;
+    auth: KeycloakJsAuthConfig | NoneAuthConfig | GetTokenAuthConfig;
     features?: FeaturesConfig;
     ui: UiConfig;
 }
