+++ ui/webpack.prod.js
@@ -3,11 +3,7 @@ const { merge } = require("webpack-merge");
 const common = require("./webpack.common.js");
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const OptimizeCSSAssetsPlugin = require("optimize-css-assets-webpack-plugin");
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-const { ModuleFederationPlugin } = require("webpack").container;
 // webpack 5 stop handling node polyfills by itself, this plugin re-enables the feature
-const NodePolyfillPlugin = require("node-polyfill-webpack-plugin")
-const { federatedModuleName, dependencies } = require("./package.json");
 
 module.exports = merge(common, {
   mode: "production",
@@ -16,37 +12,6 @@ module.exports = merge(common, {
     new MiniCssExtractPlugin({
       filename: "[name].css",
       chunkFilename: "[id].css"
-    }),
-    new NodePolyfillPlugin(),
-    new HtmlWebpackPlugin({
-      template: "./src/index.html"
-    }),
-    new ModuleFederationPlugin({
-      name: federatedModuleName,
-      filename: "remoteEntry.js",
-      exposes: {
-        "./FederatedArtifactsPage": "./src/app/pages/artifacts/artifacts.federated",
-        "./FederatedArtifactRedirectPage": "./src/app/pages/artifact/artifact.federated",
-        "./FederatedArtifactVersionPage": "./src/app/pages/artifactVersion/artifactVersion.federated",
-        "./FederatedRulesPage": "./src/app/pages/rules/rules.federated"
-      },
-      shared: {
-        ...dependencies,
-        react: {
-          eager: true,
-          singleton: true,
-          requiredVersion: dependencies["react"],
-        },
-        "react-dom": {
-          eager: true,
-          singleton: true,
-          requiredVersion: dependencies["react-dom"],
-        },
-        "react-router-dom": {
-          singleton: true,
-          requiredVersion: dependencies["react-router-dom"],
-        },
-      }
     })
   ],
   module: {
